done: Add next.config.js file to use CSS modules package, move components into their own folders, create CSS module files for each component and place each in their respective component folder
