+++ next.config.js
@@ -1,0 +1,2 @@
+const withCSS = require('@zeit/next-css')
+module.exports = withCSS({/* my next config */})
