--- src/components/Card.js
@@ -1,20 +1,0 @@
-/* eslint-disable react/react-in-jsx-scope */
-// Card
-
-export default function Card (props) {
-  const type = props.type
-
-  if (type === 'tweet') {
-    return (
-      <div className="tweet-card">
-        {props.children}
-      </div>
-    )
-  } else {
-    return (
-      <div className="card">
-        {props.children}
-      </div>
-    )
-  }
-}
