+++ src/components/Card/Card.js
@@ -1,16 +1,21 @@
-<div className="navbbar">
-    <div className="responsive">
-        menuIcon
-        logo
-        searchIcon
-    </div>
-    <div className="main">
-        logo
-        links
-    </div>
-</div>
+/* eslint-disable react/react-in-jsx-scope */
+// Card
+import styles from './Card.module.css'
 
-less than 500
-.navbbar .responsive {
-    display: true
+export default function Card (props) {
+  const type = props.type
+
+  if (type === 'tweet') {
+    return (
+      <div className="tweet-card">
+        {props.children}
+      </div>
+    )
+  } else {
+    return (
+      <div className="card">
+        {props.children}
+      </div>
+    )
+  }
 }
