+++ src/components/Card/navbar example.js
@@ -1,0 +1,16 @@
+<div className="navbbar">
+    <div className="responsive">
+        menuIcon
+        logo
+        searchIcon
+    </div>
+    <div className="main">
+        logo
+        links
+    </div>
+</div>
+
+less than 500
+.navbbar .responsive {
+    display: true
+}
