+++ src/components/Headlines/Headlines.js
@@ -4,7 +4,7 @@
 import { useState } from 'react'
 import Link from 'next/link'
 import Head from 'next/head'
-
+import styles from './Headlines.module.css'
 import { library } from '@fortawesome/fontawesome-svg-core'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 // import your icons
