+++ src/components/Headlines/Headlines.module.css
@@ -1,0 +1,76 @@
+.headlines{
+    /* margin-left: 30px;
+    margin-right: 30px; */
+  }
+  
+  .main-card{
+    width: 600px;
+  }
+  
+  .slide-control{
+    width: 20px;
+    height: 90px;
+    z-index: 3;
+    position: absolute;
+    right: 100;
+    padding-left: 575px;
+    padding-top: 10px;
+  }
+  
+  .slide-control-dots{
+    width: 12px;
+    height: 12px;
+    border-radius: 50%;
+    background-color: #E84393;
+    margin-bottom: 10px;
+    z-index: 3;
+  }
+  
+  .headline-pic{
+    width: 600px;
+    height:300px;
+    border-top-left-radius: 7px;
+    border-top-right-radius: 7px;
+  }
+  
+  .headline-card{
+    background-color: white;
+    width: 600px;
+    height: 120px;
+    padding: 10px 10px 0 10px;
+    border-bottom-left-radius: 7px;
+    border-bottom-right-radius: 7px;
+  }
+  
+  .headline-tag{
+    width: 100px;
+    background-color: #ffe7e7;
+    color: #e91e63;
+    text-align: center;
+    font-size: .8rem;
+    font-weight: bold;
+    border: 1px solid #e91e63;
+    border-radius: 25px;
+    margin: 2px 10px 10px 10px;
+  }
+  
+  .headline-news{
+    font-weight: bold;
+  }
+  
+  .social-stats{
+    float: right;
+    padding-top: 20px;
+  }
+  
+  .social-stats-icons{
+    width: 30px;
+    height: 30px;
+    margin: 0 5px 0 5px;
+  }
+  
+  .social-stat-numbers{
+    font-weight: bold;
+    padding-right: 5px;
+    padding-bottom: 2px;
+  }
