+++ src/components/MoreNews/MoreNews.js
@@ -3,7 +3,7 @@
 /* eslint-disable react/react-in-jsx-scope */
 import Link from 'next/link'
 import Head from 'next/head'
-
+import styles from './MoreNews.module.css'
 import { library } from '@fortawesome/fontawesome-svg-core'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 // import your icons
