+++ src/components/Navbar/Navbar.module.css
@@ -1,0 +1,75 @@
+.navbar{
+    width: 100%;
+    height: 69px;
+    background-color: white;
+    padding: 0 20px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+  
+  .navbar-container{
+    width: 86%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+  
+  .logo{
+    color: '#E84393';
+    margin: 0;
+  }
+  
+  .menu{
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin-left: auto;
+  }
+  
+  .menu a{
+    font-weight: bold;
+    margin-left: 20px;
+    margin-right: 20px;
+  }
+  
+  .menu a:hover{
+    border-bottom: 5px solid #E84393;
+  }
+  
+  .links{
+    display: inline-flex;
+    justify-content: center;
+    align-items: center;
+  }
+  
+  .search-area{
+    width: 50px;
+    padding-top: 25px;
+  }
+  
+  .search-bar{
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 33px;
+    padding-right: 10px;
+    background-color: #F2F2F2;
+    border-radius: 7px;
+  }
+  
+  .search-bar input{
+    height: 33px;
+    width: 172px;
+    border: 0;
+    border-radius: 7px;
+    background-color: #F2F2F2;
+  }
+  
+  .sicon{
+    /* position: absolute; */
+    width: 20px;
+    height: 20px;
+    /* top: 26px;
+    right: 165px; */
+  }
