+++ src/components/Tweets/TweetCard.js
@@ -3,7 +3,7 @@
 /* eslint-disable react/prop-types */
 import Link from 'next/link'
 import Head from 'next/head'
-import Card from './Card'
+import Card from '../Card/Card'
 
 export default function TweetCard (props) {
   return (
