+++ src/components/Tweets/Tweets.js
@@ -3,7 +3,7 @@
 import Link from 'next/link'
 import Head from 'next/head'
 import TweetCard from './TweetCard'
-
+import styles from './Tweets.module.css'
 import { library } from '@fortawesome/fontawesome-svg-core'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 // import your icons
