+++ src/components/Tweets/Tweets.module.css
@@ -1,0 +1,57 @@
+.tweets{
+
+}
+
+.tweets div{
+  margin-top: 0;
+}
+
+.tweets-group{
+  display: inline-flex;
+}
+
+
+.tweet-card{
+  background-color: white;
+  border-radius: 12px;
+  height: 90px;
+  width: 240px;
+  margin: 0 20px 0 0;
+  display: flex;
+}
+
+
+
+.tweet-profile-pic{
+  height: 65px;
+  width: 65px;
+  border-radius: 7px;
+  padding: 10px 0px 0 10px;
+}
+
+.tweet{
+  float: right;
+}
+
+.tweet-user{
+
+}
+
+.twitter-username{
+  font-size: .85rem;
+  font-weight: bold;
+  padding-left: 10px;
+  padding-right: 10px;
+}
+
+.twitter-handle{
+  font-size: .85rem;
+  color: #E84393;
+}
+
+.tweet-body{
+  color: #909090;
+  font-size: .7rem;
+  margin: 0;
+  padding: 0 10px;
+}
