+++ src/styles/globals.css
@@ -46,106 +46,11 @@ h1{
 
 
 
-.navbar{
-  width: 100%;
-  height: 69px;
-  background-color: white;
-  padding: 0 20px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
 
-.navbar-container{
-  width: 86%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
 
-.logo{
-  color: '#E84393';
-  margin: 0;
-}
 
-.menu{
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-left: auto;
-}
 
-.menu a{
-  font-weight: bold;
-  margin-left: 20px;
-  margin-right: 20px;
-}
 
-.menu a:hover{
-  border-bottom: 5px solid #E84393;
-}
-
-.links{
-  display: inline-flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.search-area{
-  width: 50px;
-  padding-top: 25px;
-}
-
-.search-bar{
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 33px;
-  padding-right: 10px;
-  background-color: #F2F2F2;
-  border-radius: 7px;
-}
-
-.search-bar input{
-  height: 33px;
-  width: 172px;
-  border: 0;
-  border-radius: 7px;
-  background-color: #F2F2F2;
-}
-
-.sicon{
-  /* position: absolute; */
-  width: 20px;
-  height: 20px;
-  /* top: 26px;
-  right: 165px; */
-}
-
-.card{
-  background-color: white;
-  border-radius: 7px;
-  height: 90px;
-  width: 215px;
-  margin-bottom: 30px;
-}
-
-.card-title{
-  font-size: 1rem;
-  padding: 10px;
-  margin-bottom: 0;
-  padding-bottom: 0;
-}
-
-.card-stats-area{
-  display: flex;
-  flex-direction: row;
-}
-
-.case-numbers{
-  margin-left: auto;
-  padding-right: 10px;
-}
 
 .flex-container{
   width: 100%;
@@ -155,57 +60,9 @@ h1{
   flex-direction: row;
 }
 
-.stats{
 
-}
 
-.filter{
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-}
 
-.stats-filter{
-  display: inline-block;
-  color: #E84393;
-  margin: 0;
-}
-
-.see-more{
-  color: #ACACAC;
-  margin: 0;
-  margin-left: auto;
-}
-
-.green-arrow-up {
-  width: 0;
-  height: 0;
-  border-left: 15px solid transparent;
-  border-right: 15px solid transparent;
-  border-bottom: 15px solid #52D57F;
-  margin-left: 25px;
-  margin-top: 25px;
-}
-
-.green-arrow-down {
-  width: 0;
-  height: 0;
-  border-left: 15px solid transparent;
-  border-right: 15px solid transparent;
-  border-top: 15px solid #52D57F;
-  margin-left: 25px;
-  margin-top: 25px;
-}
-
-.red-arrow-up {
-  width: 0;
-  height: 0;
-  border-left: 15px solid transparent;
-  border-right: 15px solid transparent;
-  border-bottom: 15px solid #DC143C;
-  margin-left: 25px;
-  margin-top: 25px;
-}
 
 .main-area{
   width: 600px;
@@ -213,188 +70,7 @@ h1{
   margin: 0 30px;
 }
 
-.headlines{
-  /* margin-left: 30px;
-  margin-right: 30px; */
-}
 
-.main-card{
-  width: 600px;
-}
-
-.slide-control{
-  width: 20px;
-  height: 90px;
-  z-index: 3;
-  position: absolute;
-  right: 100;
-  padding-left: 575px;
-  padding-top: 10px;
-}
-
-.slide-control-dots{
-  width: 12px;
-  height: 12px;
-  border-radius: 50%;
-  background-color: #E84393;
-  margin-bottom: 10px;
-  z-index: 3;
-}
-
-.headline-pic{
-  width: 600px;
-  height:300px;
-  border-top-left-radius: 7px;
-  border-top-right-radius: 7px;
-}
-
-.headline-card{
-  background-color: white;
-  width: 600px;
-  height: 120px;
-  padding: 10px 10px 0 10px;
-  border-bottom-left-radius: 7px;
-  border-bottom-right-radius: 7px;
-}
-
-.headline-tag{
-  width: 100px;
-  background-color: #ffe7e7;
-  color: #e91e63;
-  text-align: center;
-  font-size: .8rem;
-  font-weight: bold;
-  border: 1px solid #e91e63;
-  border-radius: 25px;
-  margin: 2px 10px 10px 10px;
-}
-
-.headline-news{
-  font-weight: bold;
-}
-
-.social-stats{
-  float: right;
-  padding-top: 20px;
-}
-
-.social-stats-icons{
-  width: 30px;
-  height: 30px;
-  margin: 0 5px 0 5px;
-}
-
-.social-stat-numbers{
-  font-weight: bold;
-  padding-right: 5px;
-  padding-bottom: 2px;
-}
-
-.more-news{
-  height: 434px;
-  width: 390px;
-}
-
-.heading{
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-}
-
-.more-news-card{
-  background-color: white;
-  border-radius: 7px;
-  height: 424px;
-  margin-bottom: 10px;
-}
-
-.more-news-item{
-  padding: 20px 10px 0 20px;
-  /* margin-bottom: 20px; */
-  display: flex;
-  justify-content: flex-start;
-}
-
-.more-news-item-img{
-  height: 80px;
-  width: 80px;
-  border-radius: 7px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.more-news-item-story{
-  width: 250px;
-  margin: 0;
-  padding-left: 30px;
-  font-weight: bold;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.more-news-item-story p{
-  margin: 0;
-}
-
-.tweets{
-
-}
-
-.tweets div{
-  margin-top: 0;
-}
-
-.tweets-group{
-  display: inline-flex;
-}
-
-
-.tweet-card{
-  background-color: white;
-  border-radius: 12px;
-  height: 90px;
-  width: 240px;
-  margin: 0 20px 0 0;
-  display: flex;
-}
-
-
-
-.tweet-profile-pic{
-  height: 65px;
-  width: 65px;
-  border-radius: 7px;
-  padding: 10px 0px 0 10px;
-}
-
-.tweet{
-  float: right;
-}
-
-.tweet-user{
-
-}
-
-.twitter-username{
-  font-size: .85rem;
-  font-weight: bold;
-  padding-left: 10px;
-  padding-right: 10px;
-}
-
-.twitter-handle{
-  font-size: .85rem;
-  color: #E84393;
-}
-
-.tweet-body{
-  color: #909090;
-  font-size: .7rem;
-  margin: 0;
-  padding: 0 10px;
-}
 
 
 @media only screen and (max-width: 416px) {
